externalise all processes to manage inclusion of core vs inclusion of tests
